+++ react/components/Tabs/Tab.js
@@ -7,16 +7,17 @@ class Tab extends Component {
   }
 
   render() {
-    const { active, label } = this.props
+    const { active, fullWidth, label } = this.props
     return (
       <button
         type="button"
         onClick={this.handleClick}
-        className={`vtex-tab__button pointer bt-0 bl-0 br-0 bw1 ${
+        className={`vtex-tab__button pointer bt-0 bl-0 br-0 bw1 ${fullWidth ? 'w-100' : ''} ${
           active
             ? 'near-black b--rebel-pink'
             : 'mid-gray b--transparent'}
-          hover-near-black fw5 fw4 v-mid relative pv5 ph4 f5 bg-transparent outline-0`}
+          hover-near-black fw5 fw4 v-mid relative pv5 ph4 f5 bg-transparent outline-0
+        `}
       >
         {label}
       </button>
@@ -26,6 +27,7 @@ class Tab extends Component {
 
 Tab.propTypes = {
   active: PropTypes.bool,
+  fullWidth: PropTypes.bool,
   children: PropTypes.node,
   label: PropTypes.any.isRequired,
   onClick: PropTypes.func,